added footer nav with routing
