+++ package.json
@@ -11,15 +11,21 @@
     "test": ""
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
     "@react-navigation/bottom-tabs": "^6.5.12",
+    "@react-navigation/drawer": "^6.6.7",
+    "@react-navigation/material-bottom-tabs": "^6.2.20",
     "@react-navigation/native": "^6.1.10",
     "@react-navigation/native-stack": "^6.9.18",
+    "@react-navigation/stack": "^6.3.21",
     "expo": "~50.0.7",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.4",
+    "react-native-gesture-handler": "^2.15.0",
     "react-native-paper": "^5.12.3",
-    "react-native-safe-area-context": "^4.8.2",
+    "react-native-reanimated": "^3.7.0",
+    "react-native-safe-area-context": "^4.9.0",
     "react-native-screens": "~3.29.0"
   },
   "devDependencies": {
