--- pages/UserAccountPage.js
@@ -1,15 +1,0 @@
-import { View, Text } from "react-native";
-import FooterNav from "../components/FooterNav";
-
-const UserAccountPage = () => {
-  return (
-    <>
-      <View>
-        <Text>This is the User page</Text>
-      </View>
-      <FooterNav />
-    </>
-  );
-};
-
-export default UserAccountPage;
